+++ pkg/traces/config_test.go
@@ -1331,6 +1331,80 @@ service:
       receivers: ["push_receiver", "jaeger"]
 `,
 		},
+		{
+			name: "OTLP receivers get include_metadata set to true by default",
+			cfg: `
+receivers:
+  otlp:
+    protocols:
+      grpc:
+      http:
+        endpoint: localhost:4318
+remote_write:
+  - endpoint: example.com:12345
+`,
+			expectedConfig: `
+receivers:
+  push_receiver:
+  otlp:
+    protocols:
+      grpc:
+        include_metadata: true
+      http:
+        include_metadata: true
+        endpoint: localhost:4318
+exporters:
+  otlp/0:
+    endpoint: example.com:12345
+    compression: gzip
+    retry_on_failure:
+      max_elapsed_time: 60s
+service:
+  pipelines:
+    traces:
+      exporters: ["otlp/0"]
+      processors: []
+      receivers: ["push_receiver", "otlp"]
+`,
+		},
+		{
+			name: "include_metadata respected for OTLP receivers",
+			cfg: `
+receivers:
+  otlp:
+    protocols:
+      grpc:
+        include_metadata: false
+      http:
+        include_metadata: false
+        endpoint: localhost:4318
+remote_write:
+  - endpoint: example.com:12345
+`,
+			expectedConfig: `
+receivers:
+  push_receiver:
+  otlp:
+    protocols:
+      grpc:
+        include_metadata: false
+      http:
+        include_metadata: false
+        endpoint: localhost:4318
+exporters:
+  otlp/0:
+    endpoint: example.com:12345
+    compression: gzip
+    retry_on_failure:
+      max_elapsed_time: 60s
+service:
+  pipelines:
+    traces:
+      exporters: ["otlp/0"]
+      processors: []
+      receivers: ["push_receiver", "otlp"]
+`,
+		},
 	}
 
 	for _, tc := range tt {